+++ pvl/README.md
@@ -35,9 +35,10 @@
 - [X] Create Code tab, Visual tab bar
 - [X] build tree function which is passed around to build a large tree for code building
 - [X] Clean up presentation
-- [ ] Fix tertiary drop error with tree building
+- [X] Fix tertiary drop error with tree building
 - [ ] Design the drop area
-- [ ] Code Output component - Recieves Tree and outputs HTML with parsley
+- [X] Code Output component - Recieves Tree and outputs HTML with parsley
+- [X] Preview Rendering
 - [ ] Output Targets (snippet, export html, view file, wysiwyg target)
 
 ## Considerations
