+++ pvl/src/less/layout-objects.less
@@ -60,6 +60,12 @@
                 color: @zesty-light-grey;
                 font-size: .8em;
                 padding-left: @defaultPadding;
+                display: block;
+                width: 70%;
+                overflow: hidden;
+                white-space: nowrap;
+                overflow: hidden;
+                text-overflow: ellipsis;
             }
             
         }
